test: extract tests to separate classes
