--- zeebe/exporters/camunda-exporter/src/test/java/io/camunda/exporter/handlers/batchoperation/BatchOperationStatusHandlerTest.java
@@ -1,756 +1,0 @@
-/*
- * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
- * one or more contributor license agreements. See the NOTICE file distributed
- * with this work for additional information regarding copyright ownership.
- * Licensed under the Camunda License 1.0. You may not use this file
- * except in compliance with the Camunda License 1.0.
- */
-package io.camunda.exporter.handlers.batchoperation;
-
-import static io.camunda.exporter.utils.ExporterUtil.map;
-import static io.camunda.zeebe.protocol.record.RecordMetadataDecoder.batchOperationReferenceNullValue;
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.mockito.Mockito.*;
-
-import io.camunda.exporter.store.BatchRequest;
-import io.camunda.webapps.schema.descriptors.template.OperationTemplate;
-import io.camunda.webapps.schema.entities.operation.OperationEntity;
-import io.camunda.webapps.schema.entities.operation.OperationState;
-import io.camunda.webapps.schema.entities.operation.OperationType;
-import io.camunda.zeebe.exporter.common.cache.ExporterEntityCache;
-import io.camunda.zeebe.exporter.common.cache.batchoperation.CachedBatchOperationEntity;
-import io.camunda.zeebe.protocol.record.ImmutableRecord;
-import io.camunda.zeebe.protocol.record.Record;
-import io.camunda.zeebe.protocol.record.RecordValue;
-import io.camunda.zeebe.protocol.record.RejectionType;
-import io.camunda.zeebe.protocol.record.ValueType;
-import io.camunda.zeebe.protocol.record.intent.HistoryDeletionIntent;
-import io.camunda.zeebe.protocol.record.intent.IncidentIntent;
-import io.camunda.zeebe.protocol.record.intent.ProcessInstanceIntent;
-import io.camunda.zeebe.protocol.record.intent.ProcessInstanceMigrationIntent;
-import io.camunda.zeebe.protocol.record.intent.ProcessInstanceModificationIntent;
-import io.camunda.zeebe.protocol.record.value.BpmnElementType;
-import io.camunda.zeebe.protocol.record.value.HistoryDeletionRecordValue;
-import io.camunda.zeebe.protocol.record.value.HistoryDeletionType;
-import io.camunda.zeebe.protocol.record.value.ImmutableHistoryDeletionRecordValue;
-import io.camunda.zeebe.protocol.record.value.ImmutableIncidentRecordValue;
-import io.camunda.zeebe.protocol.record.value.ImmutableProcessInstanceMigrationRecordValue;
-import io.camunda.zeebe.protocol.record.value.ImmutableProcessInstanceModificationRecordValue;
-import io.camunda.zeebe.protocol.record.value.ImmutableProcessInstanceRecordValue;
-import io.camunda.zeebe.protocol.record.value.IncidentRecordValue;
-import io.camunda.zeebe.protocol.record.value.ProcessInstanceMigrationRecordValue;
-import io.camunda.zeebe.protocol.record.value.ProcessInstanceModificationRecordValue;
-import io.camunda.zeebe.protocol.record.value.ProcessInstanceRecordValue;
-import io.camunda.zeebe.test.broker.protocol.ProtocolFactory;
-import java.time.OffsetDateTime;
-import java.util.Arrays;
-import java.util.Map;
-import java.util.Optional;
-import java.util.function.Function;
-import java.util.stream.Stream;
-import org.junit.jupiter.api.Nested;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.MethodSource;
-import org.mockito.Mockito;
-
-class BatchOperationStatusHandlerTest {
-
-  private final ProtocolFactory factory = new ProtocolFactory();
-  private final String indexName = "test-index";
-  private final long batchOperationKey = 42L;
-
-  private final ExporterEntityCache<String, CachedBatchOperationEntity> batchOperationCache =
-      mock(ExporterEntityCache.class);
-
-  abstract class AbstractOperationStatusHandlerTest<T extends RecordValue> {
-
-    final AbstractOperationStatusHandler<T> handler;
-
-    AbstractOperationStatusHandlerTest(final AbstractOperationStatusHandler<T> handler) {
-      this.handler = handler;
-
-      Mockito.reset(batchOperationCache);
-      when(batchOperationCache.get(Mockito.anyString()))
-          .thenReturn(
-              Optional.of(
-                  new CachedBatchOperationEntity(
-                      String.valueOf(batchOperationKey), map(handler.getRelevantOperationType()))));
-    }
-
-    @Test
-    void shouldHandleSuccessRecord() {
-      final var record = createSuccessRecord();
-
-      assertThat(handler.handlesRecord(record)).isTrue();
-    }
-
-    @Test
-    void shouldNotHandleSuccessRecordWithoutBatchOperationKeyInMetadata() {
-      final var record =
-          ImmutableRecord.<T>builder()
-              .from(createSuccessRecord())
-              .withBatchOperationReference(batchOperationReferenceNullValue())
-              .build();
-
-      assertThat(handler.handlesRecord(record)).isFalse();
-    }
-
-    @Test
-    void shouldNotHandleSuccessRecordWhenNotRelevantType() {
-      Mockito.reset(batchOperationCache);
-      final var otherOperationType =
-          Arrays.stream(OperationType.values())
-              .filter(t -> !t.equals(handler.getRelevantOperationType()))
-              .findAny()
-              .get();
-
-      when(batchOperationCache.get(Mockito.anyString()))
-          .thenReturn(
-              Optional.of(
-                  new CachedBatchOperationEntity(
-                      String.valueOf(batchOperationKey), map(otherOperationType))));
-
-      final var record =
-          ImmutableRecord.<T>builder()
-              .from(createSuccessRecord())
-              .withBatchOperationReference(batchOperationKey)
-              .build();
-
-      assertThat(handler.handlesRecord(record)).isFalse();
-    }
-
-    @Test
-    void shouldHandleFailureRecord() {
-      final var record = createFailureRecord();
-
-      assertThat(handler.handlesRecord(record)).isTrue();
-    }
-
-    @Test
-    void shouldNotHandleFailureRecordWithoutBatchOperationKeyInMetadata() {
-      final var record =
-          ImmutableRecord.<T>builder()
-              .from(createFailureRecord())
-              .withBatchOperationReference(batchOperationReferenceNullValue())
-              .build();
-
-      assertThat(handler.handlesRecord(record)).isFalse();
-    }
-
-    @Test
-    void shouldNotHandleFailureRecordWhenNotRelevantType() {
-      Mockito.reset(batchOperationCache);
-      final var otherOperationType =
-          Arrays.stream(OperationType.values())
-              .filter(t -> !t.equals(handler.getRelevantOperationType()))
-              .findAny()
-              .get();
-
-      when(batchOperationCache.get(Mockito.anyString()))
-          .thenReturn(
-              Optional.of(
-                  new CachedBatchOperationEntity(
-                      String.valueOf(batchOperationKey), map(otherOperationType))));
-
-      final var record =
-          ImmutableRecord.<T>builder()
-              .from(createFailureRecord())
-              .withBatchOperationReference(batchOperationKey)
-              .build();
-
-      assertThat(handler.handlesRecord(record)).isFalse();
-    }
-
-    @Test
-    void shouldNotHandleOtherRecord() {
-      final var record = factory.generateRecord(ValueType.VARIABLE);
-      assertThat(handler.handlesRecord((Record<T>) record)).isFalse();
-    }
-
-    @Test
-    void shouldUpdateEntityOnSuccess() {
-      final var record = createSuccessRecord();
-      final var entity = new OperationEntity();
-
-      handler.updateEntity(record, entity);
-
-      assertThat(entity.getState()).isEqualTo(OperationState.COMPLETED);
-      assertThat(entity.getCompletedDate()).isNotNull();
-      assertThat(entity.getErrorMessage()).isNull();
-    }
-
-    @Test
-    void shouldUpdateEntityOnFailure() {
-      final var record = createFailureRecord();
-      final var entity = new OperationEntity();
-
-      handler.updateEntity(record, entity);
-
-      assertThat(entity.getState()).isEqualTo(OperationState.FAILED);
-      assertThat(entity.getErrorMessage())
-          .isEqualTo(record.getRejectionType() + ": " + record.getRejectionReason());
-      assertThat(entity.getCompletedDate()).isNull();
-    }
-
-    @Test
-    void shouldUpdateEntityOnNotFound() {
-      final var record =
-          ImmutableRecord.<T>builder()
-              .from(createFailureRecord())
-              .withRejectionType(RejectionType.NOT_FOUND)
-              .build();
-
-      final var entity = new OperationEntity();
-
-      handler.updateEntity(record, entity);
-
-      assertThat(entity.getState()).isEqualTo(OperationState.SKIPPED);
-      assertThat(entity.getErrorMessage()).isNull();
-      assertThat(entity.getCompletedDate()).isNull();
-    }
-
-    @Test
-    void shouldFlushEntityFields() {
-      final var entity = new OperationEntity();
-      entity.setState(OperationState.COMPLETED);
-      entity.setType(handler.getRelevantOperationType());
-      entity.setBatchOperationId(String.valueOf(batchOperationKey));
-      entity.setItemKey(123L);
-      entity.setProcessInstanceKey(456L);
-      entity.setCompletedDate(OffsetDateTime.now());
-      entity.setErrorMessage("error message");
-
-      final var mockRequest = mock(BatchRequest.class);
-
-      // when
-      handler.flush(entity, mockRequest);
-
-      // then
-      verify(mockRequest, times(1))
-          .upsert(
-              indexName,
-              entity.getId(),
-              entity,
-              Map.of(
-                  OperationTemplate.STATE, entity.getState(),
-                  OperationTemplate.TYPE, entity.getType(),
-                  OperationTemplate.BATCH_OPERATION_ID, entity.getBatchOperationId(),
-                  OperationTemplate.ITEM_KEY, entity.getItemKey(),
-                  OperationTemplate.PROCESS_INSTANCE_KEY, entity.getProcessInstanceKey(),
-                  OperationTemplate.COMPLETED_DATE, entity.getCompletedDate(),
-                  OperationTemplate.ERROR_MSG, entity.getErrorMessage()));
-    }
-
-    @Test
-    void shouldGenerateCorrectId() {
-      final var record = createSuccessRecord();
-      final var generatedIds = handler.generateIds(record);
-
-      assertThat(generatedIds)
-          .containsExactly(batchOperationKey + "_" + handler.getItemKey(record));
-    }
-
-    @ParameterizedTest
-    @MethodSource("recordProviderStream")
-    void shouldUpdateEntitySetRootProcessInstanceKey(
-        final Function<AbstractOperationStatusHandlerTest<T>, Record<T>> recordProvider) {
-      final var record = recordProvider.apply(this);
-      final var entity = new OperationEntity();
-
-      handler.updateEntity(record, entity);
-
-      if (shouldHandleRootProcessInstanceKey()) {
-        assertThat(entity.getRootProcessInstanceKey())
-            .isPositive()
-            .isEqualTo(getRootProcessInstanceKey(record));
-      } else {
-        assertThat(entity.getRootProcessInstanceKey()).isNull();
-      }
-    }
-
-    static <T extends RecordValue>
-        Stream<Function<AbstractOperationStatusHandlerTest<T>, Record<T>>> recordProviderStream() {
-      return Stream.of(
-          AbstractOperationStatusHandlerTest::createSuccessRecord,
-          AbstractOperationStatusHandlerTest::createFailureRecord);
-    }
-
-    @Test
-    void shouldExtractCorrectItemKeyTest() {
-      shouldExtractCorrectItemKey();
-    }
-
-    abstract void shouldExtractCorrectItemKey();
-
-    @Test
-    void shouldExtractCorrectProcessInstanceKeyTest() {
-      shouldExtractCorrectProcessInstanceKey();
-    }
-
-    abstract void shouldExtractCorrectProcessInstanceKey();
-
-    abstract Record<T> createSuccessRecord();
-
-    abstract Record<T> createFailureRecord();
-
-    long getRootProcessInstanceKey(final Record<T> record) {
-      return handler.getRootProcessInstanceKey(record);
-    }
-
-    boolean shouldHandleRootProcessInstanceKey() {
-      return true;
-    }
-  }
-
-  @Nested
-  class ProcessInstanceModificationOperationHandlerTest
-      extends AbstractOperationStatusHandlerTest<ProcessInstanceModificationRecordValue> {
-
-    ProcessInstanceModificationOperationHandlerTest() {
-      super(new ProcessInstanceModificationOperationHandler(indexName, batchOperationCache));
-    }
-
-    @Override
-    void shouldExtractCorrectItemKey() {
-      final var record = createSuccessRecord();
-      final var itemKey = handler.getItemKey(record);
-
-      assertThat(itemKey).isEqualTo(record.getValue().getProcessInstanceKey());
-    }
-
-    @Override
-    void shouldExtractCorrectProcessInstanceKey() {
-      final var record = createSuccessRecord();
-      final var processInstanceKey = handler.getProcessInstanceKey(record);
-
-      assertThat(processInstanceKey).isEqualTo(record.getValue().getProcessInstanceKey());
-    }
-
-    @Override
-    Record<ProcessInstanceModificationRecordValue> createSuccessRecord() {
-      return factory.generateRecord(
-          ValueType.PROCESS_INSTANCE_MODIFICATION,
-          b ->
-              b.withIntent(ProcessInstanceModificationIntent.MODIFIED)
-                  .withBatchOperationReference(batchOperationKey));
-    }
-
-    @Override
-    Record<ProcessInstanceModificationRecordValue> createFailureRecord() {
-      return factory.generateRecord(
-          ValueType.PROCESS_INSTANCE_MODIFICATION,
-          b ->
-              b.withRejectionType(RejectionType.PROCESSING_ERROR)
-                  .withIntent(ProcessInstanceModificationIntent.MODIFY)
-                  .withBatchOperationReference(batchOperationKey));
-    }
-
-    @Test
-    void shouldNotSetRootProcessInstanceKeyWhenDefault() {
-      // given
-      final Record<ProcessInstanceModificationRecordValue> record =
-          factory.generateRecord(
-              ValueType.PROCESS_INSTANCE_MODIFICATION,
-              b ->
-                  b.withRejectionType(RejectionType.PROCESSING_ERROR)
-                      .withIntent(ProcessInstanceModificationIntent.MODIFY)
-                      .withValue(
-                          ImmutableProcessInstanceModificationRecordValue.builder()
-                              .from(
-                                  factory.generateObject(
-                                      ProcessInstanceModificationRecordValue.class))
-                              .withRootProcessInstanceKey(-1L)
-                              .build())
-                      .withBatchOperationReference(batchOperationKey));
-
-      // when
-      final var entity = new OperationEntity();
-      handler.updateEntity(record, entity);
-
-      // then
-      assertThat(entity.getRootProcessInstanceKey()).isNull();
-    }
-  }
-
-  @Nested
-  class ProcessInstanceMigrationOperationHandlerTest
-      extends AbstractOperationStatusHandlerTest<ProcessInstanceMigrationRecordValue> {
-
-    ProcessInstanceMigrationOperationHandlerTest() {
-      super(new ProcessInstanceMigrationOperationHandler(indexName, batchOperationCache));
-    }
-
-    @Override
-    void shouldExtractCorrectItemKey() {
-      final var record = createSuccessRecord();
-      final var itemKey = handler.getItemKey(record);
-
-      assertThat(itemKey).isEqualTo(record.getValue().getProcessInstanceKey());
-    }
-
-    @Override
-    void shouldExtractCorrectProcessInstanceKey() {
-      final var record = createSuccessRecord();
-      final var processInstanceKey = handler.getProcessInstanceKey(record);
-
-      assertThat(processInstanceKey).isEqualTo(record.getValue().getProcessInstanceKey());
-    }
-
-    @Override
-    Record<ProcessInstanceMigrationRecordValue> createSuccessRecord() {
-      return factory.generateRecord(
-          ValueType.PROCESS_INSTANCE_MIGRATION,
-          b ->
-              b.withIntent(ProcessInstanceMigrationIntent.MIGRATED)
-                  .withBatchOperationReference(batchOperationKey));
-    }
-
-    @Override
-    Record<ProcessInstanceMigrationRecordValue> createFailureRecord() {
-      return factory.generateRecord(
-          ValueType.PROCESS_INSTANCE_MIGRATION,
-          b ->
-              b.withRejectionType(RejectionType.PROCESSING_ERROR)
-                  .withIntent(ProcessInstanceMigrationIntent.MIGRATE)
-                  .withBatchOperationReference(batchOperationKey));
-    }
-
-    @Test
-    void shouldNotSetRootProcessInstanceKeyWhenDefault() {
-      // given
-      final Record<ProcessInstanceMigrationRecordValue> record =
-          factory.generateRecord(
-              ValueType.PROCESS_INSTANCE_MIGRATION,
-              b ->
-                  b.withRejectionType(RejectionType.PROCESSING_ERROR)
-                      .withIntent(ProcessInstanceMigrationIntent.MIGRATE)
-                      .withValue(
-                          ImmutableProcessInstanceMigrationRecordValue.builder()
-                              .from(
-                                  factory.generateObject(ProcessInstanceMigrationRecordValue.class))
-                              .withRootProcessInstanceKey(-1L)
-                              .build())
-                      .withBatchOperationReference(batchOperationKey));
-
-      // when
-      final var entity = new OperationEntity();
-      handler.updateEntity(record, entity);
-
-      // then
-      assertThat(entity.getRootProcessInstanceKey()).isNull();
-    }
-  }
-
-  @Nested
-  class ProcessInstanceCancellationOperationHandlerTest
-      extends AbstractOperationStatusHandlerTest<ProcessInstanceRecordValue> {
-
-    ProcessInstanceCancellationOperationHandlerTest() {
-      super(new ProcessInstanceCancellationOperationHandler(indexName, batchOperationCache));
-    }
-
-    @Test
-    void shouldNotHandleRecordOfSubprocess() {
-      final Record<ProcessInstanceRecordValue> record =
-          factory.generateRecord(
-              ValueType.PROCESS_INSTANCE,
-              b ->
-                  b.withIntent(ProcessInstanceIntent.ELEMENT_TERMINATED)
-                      .withValue(
-                          ImmutableProcessInstanceRecordValue.builder()
-                              .from(factory.generateObject(ProcessInstanceRecordValue.class))
-                              .withBpmnElementType(BpmnElementType.PROCESS)
-                              .build())
-                      .withBatchOperationReference(batchOperationKey));
-
-      assertThat(handler.handlesRecord(record)).isFalse();
-    }
-
-    @Override
-    void shouldExtractCorrectItemKey() {
-      final var record = createSuccessRecord();
-      final var itemKey = handler.getItemKey(record);
-
-      assertThat(itemKey).isEqualTo(record.getValue().getProcessInstanceKey());
-    }
-
-    @Override
-    void shouldExtractCorrectProcessInstanceKey() {
-      final var record = createSuccessRecord();
-      final var processInstanceKey = handler.getProcessInstanceKey(record);
-
-      assertThat(processInstanceKey).isEqualTo(record.getValue().getProcessInstanceKey());
-    }
-
-    @Override
-    Record<ProcessInstanceRecordValue> createSuccessRecord() {
-      return factory.generateRecord(
-          ValueType.PROCESS_INSTANCE,
-          b ->
-              b.withIntent(ProcessInstanceIntent.ELEMENT_TERMINATED)
-                  .withValue(
-                      ImmutableProcessInstanceRecordValue.builder()
-                          .from(factory.generateObject(ProcessInstanceRecordValue.class))
-                          .withBpmnElementType(BpmnElementType.PROCESS)
-                          .withParentProcessInstanceKey(-1L)
-                          .build())
-                  .withBatchOperationReference(batchOperationKey));
-    }
-
-    @Override
-    Record<ProcessInstanceRecordValue> createFailureRecord() {
-      return factory.generateRecord(
-          ValueType.PROCESS_INSTANCE_MIGRATION,
-          b ->
-              b.withRejectionType(RejectionType.PROCESSING_ERROR)
-                  .withValue(
-                      ImmutableProcessInstanceRecordValue.builder()
-                          .from(factory.generateObject(ProcessInstanceRecordValue.class))
-                          .withBpmnElementType(BpmnElementType.PROCESS)
-                          .withParentProcessInstanceKey(-1L)
-                          .build())
-                  .withIntent(ProcessInstanceIntent.CANCEL)
-                  .withBatchOperationReference(batchOperationKey));
-    }
-
-    @Test
-    void shouldNotSetRootProcessInstanceKeyWhenDefault() {
-      // given
-      final Record<ProcessInstanceRecordValue> record =
-          factory.generateRecord(
-              ValueType.PROCESS_INSTANCE,
-              b ->
-                  b.withValue(
-                          ImmutableProcessInstanceRecordValue.builder()
-                              .from(factory.generateObject(ProcessInstanceRecordValue.class))
-                              .withRootProcessInstanceKey(-1L)
-                              .build())
-                      .withBatchOperationReference(batchOperationKey));
-
-      // when
-      final var entity = new OperationEntity();
-      handler.updateEntity(record, entity);
-
-      // then
-      assertThat(entity.getRootProcessInstanceKey()).isNull();
-    }
-  }
-
-  @Nested
-  class ResolveIncidentOperationHandlerTest
-      extends AbstractOperationStatusHandlerTest<IncidentRecordValue> {
-
-    ResolveIncidentOperationHandlerTest() {
-      super(new ResolveIncidentOperationHandler(indexName, batchOperationCache));
-    }
-
-    @Override
-    void shouldExtractCorrectItemKey() {
-      final var record = createSuccessRecord();
-      final var itemKey = handler.getItemKey(record);
-
-      assertThat(itemKey).isEqualTo(record.getKey());
-    }
-
-    @Override
-    void shouldExtractCorrectProcessInstanceKey() {
-      final var record = createSuccessRecord();
-      final var processInstanceKey = handler.getProcessInstanceKey(record);
-
-      assertThat(processInstanceKey).isEqualTo(record.getValue().getProcessInstanceKey());
-    }
-
-    @Override
-    Record<IncidentRecordValue> createSuccessRecord() {
-      return factory.generateRecord(
-          ValueType.INCIDENT,
-          b ->
-              b.withIntent(IncidentIntent.RESOLVED).withBatchOperationReference(batchOperationKey));
-    }
-
-    @Override
-    Record<IncidentRecordValue> createFailureRecord() {
-      return factory.generateRecord(
-          ValueType.INCIDENT,
-          b ->
-              b.withRejectionType(RejectionType.PROCESSING_ERROR)
-                  .withIntent(IncidentIntent.RESOLVE)
-                  .withBatchOperationReference(batchOperationKey));
-    }
-
-    @Test
-    void shouldNotSetRootProcessInstanceKeyWhenDefault() {
-      // given
-      final Record<IncidentRecordValue> record =
-          factory.generateRecord(
-              ValueType.INCIDENT,
-              b ->
-                  b.withValue(
-                          ImmutableIncidentRecordValue.builder()
-                              .from(factory.generateObject(IncidentRecordValue.class))
-                              .withRootProcessInstanceKey(-1L)
-                              .build())
-                      .withBatchOperationReference(batchOperationKey));
-
-      // when
-      final var entity = new OperationEntity();
-      handler.updateEntity(record, entity);
-
-      // then
-      assertThat(entity.getRootProcessInstanceKey()).isNull();
-    }
-  }
-
-  @Nested
-  class ProcessInstanceHistoryDeletionOperationHandlerTest
-      extends AbstractOperationStatusHandlerTest<HistoryDeletionRecordValue> {
-
-    ProcessInstanceHistoryDeletionOperationHandlerTest() {
-      super(new ProcessInstanceHistoryDeletionOperationHandler(indexName, batchOperationCache));
-    }
-
-    @Override
-    @Test
-    void shouldHandleSuccessRecord() {
-      // We override the abstract test as this handler should not handle a success record.
-      final var record = createSuccessRecord();
-      assertThat(handler.handlesRecord(record)).isFalse();
-    }
-
-    @Override
-    @Test
-    void shouldUpdateEntityOnSuccess() {
-      // We don't want to test anything here. This handler is not supposed to handle the success
-      // record, and thus should not update the entity. Since the abstract test would fail, we need
-      // to override it with a NOOP implementation.
-      assert true;
-    }
-
-    @Override
-    void shouldExtractCorrectItemKey() {
-      final var record = createSuccessRecord();
-      final var itemKey = handler.getItemKey(record);
-
-      assertThat(itemKey).isEqualTo(record.getValue().getResourceKey());
-    }
-
-    @Override
-    void shouldExtractCorrectProcessInstanceKey() {
-      final var record = createSuccessRecord();
-      final var processInstanceKey = handler.getProcessInstanceKey(record);
-
-      assertThat(processInstanceKey).isEqualTo(record.getValue().getResourceKey());
-    }
-
-    @Override
-    Record<HistoryDeletionRecordValue> createSuccessRecord() {
-      return factory.generateRecord(
-          ValueType.HISTORY_DELETION,
-          b ->
-              b.withIntent(HistoryDeletionIntent.DELETED)
-                  .withBatchOperationReference(batchOperationKey)
-                  .withValue(
-                      ImmutableHistoryDeletionRecordValue.builder()
-                          .from(factory.generateObject(HistoryDeletionRecordValue.class))
-                          .withResourceType(HistoryDeletionType.PROCESS_INSTANCE)
-                          .build()));
-    }
-
-    @Override
-    Record<HistoryDeletionRecordValue> createFailureRecord() {
-      return factory.generateRecord(
-          ValueType.HISTORY_DELETION,
-          b ->
-              b.withRejectionType(RejectionType.PROCESSING_ERROR)
-                  .withIntent(HistoryDeletionIntent.DELETE)
-                  .withBatchOperationReference(batchOperationKey)
-                  .withValue(
-                      ImmutableHistoryDeletionRecordValue.builder()
-                          .from(factory.generateObject(HistoryDeletionRecordValue.class))
-                          .withResourceType(HistoryDeletionType.PROCESS_INSTANCE)
-                          .build()));
-    }
-
-    @Override
-    boolean shouldHandleRootProcessInstanceKey() {
-      return false;
-    }
-  }
-
-  @Nested
-  class DecisionInstanceHistoryDeletionOperationHandlerTest
-      extends AbstractOperationStatusHandlerTest<HistoryDeletionRecordValue> {
-
-    DecisionInstanceHistoryDeletionOperationHandlerTest() {
-      super(new DecisionInstanceHistoryDeletionOperationHandler(indexName, batchOperationCache));
-    }
-
-    @Override
-    @Test
-    void shouldHandleSuccessRecord() {
-      // We override the abstract test as this handler should not handle a success record.
-      final var record = createSuccessRecord();
-      assertThat(handler.handlesRecord(record)).isFalse();
-    }
-
-    @Override
-    @Test
-    void shouldUpdateEntityOnSuccess() {
-      // We don't want to test anything here. This handler is not supposed to handle the success
-      // record, and thus should not update the entity. Since the abstract test would fail, we need
-      // to override it with a NOOP implementation.
-      assert true;
-    }
-
-    @Override
-    void shouldExtractCorrectItemKey() {
-      final var record = createSuccessRecord();
-      final var itemKey = handler.getItemKey(record);
-
-      assertThat(itemKey).isEqualTo(record.getValue().getResourceKey());
-    }
-
-    @Override
-    void shouldExtractCorrectProcessInstanceKey() {
-      final var record = createSuccessRecord();
-      final var processInstanceKey = handler.getProcessInstanceKey(record);
-
-      assertThat(processInstanceKey).isEqualTo(-1L);
-    }
-
-    @Override
-    Record<HistoryDeletionRecordValue> createSuccessRecord() {
-      return factory.generateRecord(
-          ValueType.HISTORY_DELETION,
-          b ->
-              b.withIntent(HistoryDeletionIntent.DELETED)
-                  .withBatchOperationReference(batchOperationKey)
-                  .withValue(
-                      ImmutableHistoryDeletionRecordValue.builder()
-                          .from(factory.generateObject(HistoryDeletionRecordValue.class))
-                          .withResourceType(HistoryDeletionType.DECISION_INSTANCE)
-                          .build()));
-    }
-
-    @Override
-    Record<HistoryDeletionRecordValue> createFailureRecord() {
-      return factory.generateRecord(
-          ValueType.HISTORY_DELETION,
-          b ->
-              b.withRejectionType(RejectionType.PROCESSING_ERROR)
-                  .withIntent(HistoryDeletionIntent.DELETE)
-                  .withBatchOperationReference(batchOperationKey)
-                  .withValue(
-                      ImmutableHistoryDeletionRecordValue.builder()
-                          .from(factory.generateObject(HistoryDeletionRecordValue.class))
-                          .withResourceType(HistoryDeletionType.DECISION_INSTANCE)
-                          .build()));
-    }
-
-    @Override
-    boolean shouldHandleRootProcessInstanceKey() {
-      return false;
-    }
-  }
-}
